feat: validate chat access
