+++ frontend/app/chat/[chatId]/hooks/useQuestion.ts
@@ -1,11 +1,10 @@
-import axios from "axios";
 import { useTranslation } from "react-i18next";
 
 import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
-import { useChatContext } from "@/lib/context/ChatProvider/hooks/useChatContext";
 import { useFetch, useToast } from "@/lib/hooks";
 
-import { ChatMessage, ChatQuestion } from "../types";
+import { useHandleStream } from "./useHandleStream";
+import { ChatQuestion } from "../types";
 
 interface UseChatService {
   addStreamQuestion: (
@@ -16,43 +15,29 @@ interface UseChatService {
 
 export const useQuestion = (): UseChatService => {
   const { fetchInstance } = useFetch();
-  const { updateStreamingHistory } = useChatContext();
   const { currentBrain } = useBrainContext();
 
   const { t } = useTranslation(["chat"]);
   const { publish } = useToast();
+  const { handleStream } = useHandleStream();
 
-  const handleStream = async (
-    reader: ReadableStreamDefaultReader<Uint8Array>
-  ): Promise<void> => {
-    const decoder = new TextDecoder("utf-8");
-
-    const handleStreamRecursively = async () => {
-      const { done, value } = await reader.read();
-
-      if (done) {
-        return;
-      }
-
-      const dataStrings = decoder
-        .decode(value)
-        .trim()
-        .split("data: ")
-        .filter(Boolean);
-
-      dataStrings.forEach((data) => {
-        try {
-          const parsedData = JSON.parse(data) as ChatMessage;
-          updateStreamingHistory(parsedData);
-        } catch (error) {
-          console.error(t("errorParsingData", { ns: "chat" }), error);
-        }
+  const handleFetchError = async (response: Response) => {
+    if (response.status === 429) {
+      publish({
+        variant: "danger",
+        text: t("tooManyRequests", { ns: "chat" }),
       });
 
-      await handleStreamRecursively();
-    };
+      return;
+    }
 
-    await handleStreamRecursively();
+    const errorMessage = (await response.json()) as { detail: string };
+    publish({
+      variant: "danger",
+      text: errorMessage.detail,
+    });
+
+    return;
   };
 
   const addStreamQuestion = async (
@@ -64,29 +49,29 @@ export const useQuestion = (): UseChatService => {
       Accept: "text/event-stream",
     };
     const body = JSON.stringify(chatQuestion);
-    console.log("Calling API...");
+
     try {
       const response = await fetchInstance.post(
         `/chat/${chatId}/question/stream?brain_id=${currentBrain?.id ?? ""}`,
         body,
         headers
       );
+      if (!response.ok) {
+        void handleFetchError(response);
+
+        return;
+      }
 
       if (response.body === null) {
         throw new Error(t("resposeBodyNull", { ns: "chat" }));
       }
 
-      console.log(t("receivedResponse"), response);
       await handleStream(response.body.getReader());
     } catch (error) {
-      if (axios.isAxiosError(error) && error.response?.status === 429) {
-        publish({
-          variant: "danger",
-          text: t("tooManyRequests", { ns: "chat" }),
-        });
-      }
-
-      console.error(t("errorCallingAPI", { ns: "chat" }), error);
+      publish({
+        variant: "danger",
+        text: String(error),
+      });
     }
   };
